+++ include/unit.h
@@ -9,13 +9,13 @@
 class Unit : public Selectable {
  protected:
     std::string name;
-    uint16_t spice;
-    uint16_t speed;
+    int spice;
+    int speed;
 
  public:
     Unit() = default;
-    Unit(uint16_t LP,uint16_t spice, Position pos, uint8_t dim_x, uint8_t dim_y,uint16_t speed);
-    bool place(Board & board,std::vector<Position> & positions,uint16_t & spice);
+    Unit(int LP,int spice, Position pos, int dim_x, int dim_y,int speed);
+    bool place(Board & board,std::vector<Position> & positions,int & spice);
     virtual void react(Cell & location);
     virtual ~Unit();
 };
@@ -24,23 +24,23 @@ class Unit : public Selectable {
 
 /*
 class Harvester : public Unit {
-   uint16_t stored_spice;
-   uint16_t max_spice;
+   int stored_spice;
+   int max_spice;
 public:
-    Harvester(uint16_t LP,uint16_t spice, Position pos, uint8_t dim_x, uint8_t dim_y,uint16_t speed,uint16_t stored_spice, uint16_t max_spice);
+    Harvester(int LP,int spice, Position pos, int dim_x, int dim_y,int speed,int stored_spice, int max_spice);
     void create();
     void react(Cell & location);
 };
 */
 class Trike : public Selectable {
    std::string name;
-   uint16_t spice;
-   uint16_t speed;
-   uint16_t attack;
+   int spice;
+   int speed;
+   int attack;
 public:
-    Trike(uint16_t LP,uint16_t spice, Position pos, uint8_t dim_x, uint8_t dim_y,uint16_t speed,uint16_t attack);
+    Trike(int LP,int spice, Position pos, int dim_x, int dim_y,int speed,int attack);
     void create();
     void react(Cell & location);
-    bool place(Board & board,std::vector<Position> & positions,uint16_t & spice);
+    bool place(Board & board,std::vector<Position> & positions,int & spice);
     ~Trike();
 };
